+++ Assets/_Project/Scripts/Water System/Editor/FloodSimDataNewEditor.cs
@@ -18,20 +18,15 @@ public class FloodSimDataEditor : Editor
         EditorGUILayout.LabelField("Simulation Area", $"{data.N} x {data.N}");
         EditorGUILayout.LabelField("Status", data.IsInitialized ? "Initialized" : "Not Initialized");
 
-        if (data.NewTerrainDataSource != null)
+        if (data.TerrainDataSource != null)
         {
-            EditorGUILayout.LabelField("New Terrain Source", data.NewTerrainDataSource.name);
-            EditorGUILayout.LabelField("New Terrain Data Loaded", data.NewTerrainDataSource.DataLoaded ? "Yes" : "No");
-            if (data.NewTerrainDataSource.DataLoaded)
+            EditorGUILayout.LabelField("Terrain Source", data.TerrainDataSource.name);
+            EditorGUILayout.LabelField("Terrain Data Loaded", data.TerrainDataSource.DataLoaded ? "Yes" : "No");
+            if (data.TerrainDataSource.DataLoaded)
             {
-                EditorGUILayout.LabelField("Elevation Range", $"[{data.NewTerrainDataSource.MinElevation}, {data.NewTerrainDataSource.MaxElevation}]");
-                EditorGUILayout.LabelField("Tiles Loaded", data.NewTerrainDataSource.TotalTilesWritten.ToString());
+                EditorGUILayout.LabelField("Elevation Range", $"[{data.TerrainDataSource.MinElevation}, {data.TerrainDataSource.MaxElevation}]");
+                EditorGUILayout.LabelField("Tiles Loaded", data.TerrainDataSource.TotalTilesWritten.ToString());
             }
-        }
-        else if (data.TerrainDataSource != null)
-        {
-            EditorGUILayout.LabelField("Old Terrain Source", data.TerrainDataSource.name);
-            EditorGUILayout.LabelField("Old Terrain Data Loaded", data.TerrainDataSource.DataLoaded ? "Yes" : "No");
         }
         else
         {
@@ -40,11 +35,6 @@ public class FloodSimDataEditor : Editor
 
         EditorGUILayout.Space();
         EditorGUILayout.HelpBox("This ScriptableObject only contains data. Use a FloodSimulationManager component to run the simulation.", MessageType.Info);
-        
-        if (data.NewTerrainDataSource != null && data.TerrainDataSource != null)
-        {
-            EditorGUILayout.HelpBox("Both terrain data sources are assigned. NewTerrainData (z-value based) will take priority over the old TerrainData (tile type based).", MessageType.Info);
-        }
     }
 }
 #endif
