--- Assets/_Project/Scripts/Water System/Editor/NewTerrainDataEditor.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3c16afa41b975af428903bcf0f179506
