--- Assets/_Project/Scripts/Water System/Editor/NewTerrainLoaderEditor.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 807a6744c5f85414cbafa3be05e6f1f1
