+++ Assets/_Project/Scripts/Water System/FloodSimulationManager.cs
@@ -64,17 +64,10 @@ public class FloodSimulationManager : MonoBehaviour
         int gridHeight = simulationData.GridHeight;
         int N = simulationData.N;
         
-        // Priority 1: Check for NewTerrainData (z-value based system)
-        if (simulationData.NewTerrainDataSource != null && simulationData.NewTerrainDataSource.DataLoaded)
-        {
-            if (InitializeWithNewTerrainData(gridWidth, gridHeight, N))
-                return;
-        }
-        
-        // Priority 2: Check for old TerrainData (tile type based system)
+        // Check for terrain data source
         if (simulationData.TerrainDataSource != null && simulationData.TerrainDataSource.DataLoaded)
         {
-            if (InitializeWithOldTerrainData(gridWidth, gridHeight, N))
+            if (InitializeWithTerrainData(gridWidth, gridHeight, N))
                 return;
         }
         
@@ -83,73 +76,9 @@ public class FloodSimulationManager : MonoBehaviour
     }
 
     /// <summary>
-    /// Initialize simulation using the new NewTerrainData (z-value based system)
+    /// Initialize simulation using the TerrainData (z-value based system)
     /// </summary>
-    private bool InitializeWithNewTerrainData(int gridWidth, int gridHeight, int N)
-    {
-        // Find a NewTerrainLoader to help us convert the data
-        NewTerrainLoader newTerrainLoader = FindObjectOfType<NewTerrainLoader>();
-        
-        if (newTerrainLoader == null)
-        {
-            Debug.LogWarning("[FloodSimulationManager] NewTerrainData source found but no NewTerrainLoader in scene to convert data");
-            return false;
-        }
-
-        // Calculate offsets to shift terrain coordinates into valid [0, N-1] range
-        int offsetX = 0, offsetY = 0;
-        if (simulationData.NewTerrainDataSource.TilePositions.Count > 0)
-        {
-            int minX = int.MaxValue, minY = int.MaxValue;
-            foreach (var pos in simulationData.NewTerrainDataSource.TilePositions)
-            {
-                minX = Mathf.Min(minX, pos.x);
-                minY = Mathf.Min(minY, pos.y);
-            }
-            // Shift negative coordinates to start at 0
-            offsetX = -minX;
-            offsetY = -minY;
-        }
-        
-        float[,] terrainFromData = newTerrainLoader.ConvertToSimulationGrid(N, N, offsetX, offsetY);
-        if (terrainFromData != null)
-        {
-            // Initialize arrays
-            simulationData.water = new float[gridWidth, gridHeight];
-            simulationData.terrain = new float[gridWidth, gridHeight];
-            simulationData.flowX = new float[gridWidth, gridHeight];
-            simulationData.flowY = new float[gridWidth, gridHeight];
-
-            // Copy terrain data from NewTerrainData (offset by 1 for boundary)
-            for (int y = 0; y < N; y++)
-            {
-                for (int x = 0; x < N; x++)
-                {
-                    simulationData.terrain[x + 1, y + 1] = terrainFromData[x, y];
-                    // Water should be a thin layer on top of terrain, not a separate height
-                    simulationData.water[x + 1, y + 1] = simulationData.startingWaterDepth;
-                }
-            }
-
-            SetupBoundaryWalls(gridWidth, gridHeight);
-            
-            simulationData.IsInitialized = true;
-            
-            Debug.Log($"[FloodSimulationManager] Initialized with NewTerrainData: {simulationData.NewTerrainDataSource.TotalTilesWritten} tiles");
-            Debug.Log($"[FloodSimulationManager] Elevation range: [{simulationData.NewTerrainDataSource.MinElevation}, {simulationData.NewTerrainDataSource.MaxElevation}]");
-            
-            OnSimulationInitialized?.Invoke();
-            OnSimulationStep?.Invoke();
-            return true;
-        }
-        
-        return false;
-    }
-
-    /// <summary>
-    /// Initialize simulation using the old TerrainData (tile type based system)
-    /// </summary>
-    private bool InitializeWithOldTerrainData(int gridWidth, int gridHeight, int N)
+    private bool InitializeWithTerrainData(int gridWidth, int gridHeight, int N)
     {
         // Find a TerrainLoader to help us convert the data
         TerrainLoader terrainLoader = FindObjectOfType<TerrainLoader>();
@@ -175,7 +104,7 @@ public class FloodSimulationManager : MonoBehaviour
             offsetY = -minY;
         }
         
-        float[,] terrainFromData = terrainLoader.ConvertToHeightArray(simulationData.TerrainDataSource, N, N, offsetX, offsetY);
+        float[,] terrainFromData = terrainLoader.ConvertToSimulationGrid(N, N, offsetX, offsetY);
         if (terrainFromData != null)
         {
             // Initialize arrays
@@ -196,10 +125,11 @@ public class FloodSimulationManager : MonoBehaviour
             }
 
             SetupBoundaryWalls(gridWidth, gridHeight);
-
+            
             simulationData.IsInitialized = true;
             
-            Debug.Log($"[FloodSimulationManager] Initialized with old TerrainData: {simulationData.TerrainDataSource.TotalTilesWritten} tiles");
+            Debug.Log($"[FloodSimulationManager] Initialized with TerrainData: {simulationData.TerrainDataSource.TotalTilesWritten} tiles");
+            Debug.Log($"[FloodSimulationManager] Elevation range: [{simulationData.TerrainDataSource.MinElevation}, {simulationData.TerrainDataSource.MaxElevation}]");
             
             OnSimulationInitialized?.Invoke();
             OnSimulationStep?.Invoke();
