--- Assets/_Project/Scripts/Water System/NewTerrainData.asset
@@ -1,121 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: de8d881e0a46a2542a9cb41813341c16, type: 3}
-  m_Name: NewTerrainData
-  m_EditorClassIdentifier: 
-  tilePositions:
-  - {x: -5, y: -6, z: 2}
-  - {x: -5, y: -5, z: 2}
-  - {x: -5, y: -4, z: 2}
-  - {x: -5, y: -3, z: 2}
-  - {x: -5, y: -2, z: 2}
-  - {x: -5, y: -1, z: 2}
-  - {x: -5, y: 0, z: 2}
-  - {x: -5, y: 1, z: 2}
-  - {x: -5, y: 2, z: 2}
-  - {x: -5, y: 3, z: 2}
-  - {x: -4, y: -6, z: 2}
-  - {x: -4, y: -5, z: 1}
-  - {x: -4, y: -4, z: 1}
-  - {x: -4, y: -3, z: 1}
-  - {x: -4, y: -2, z: 1}
-  - {x: -4, y: -1, z: 1}
-  - {x: -4, y: 0, z: 1}
-  - {x: -4, y: 1, z: 1}
-  - {x: -4, y: 2, z: 1}
-  - {x: -4, y: 3, z: 2}
-  - {x: -3, y: -6, z: 2}
-  - {x: -3, y: -5, z: 1}
-  - {x: -3, y: -4, z: 0}
-  - {x: -3, y: -3, z: 0}
-  - {x: -3, y: -2, z: 0}
-  - {x: -3, y: -1, z: 0}
-  - {x: -3, y: 0, z: 0}
-  - {x: -3, y: 1, z: 0}
-  - {x: -3, y: 2, z: 1}
-  - {x: -3, y: 3, z: 2}
-  - {x: -2, y: -6, z: 2}
-  - {x: -2, y: -5, z: 1}
-  - {x: -2, y: -4, z: 0}
-  - {x: -2, y: -3, z: 0}
-  - {x: -2, y: -2, z: 0}
-  - {x: -2, y: -1, z: 0}
-  - {x: -2, y: 0, z: 0}
-  - {x: -2, y: 1, z: 0}
-  - {x: -2, y: 2, z: 1}
-  - {x: -2, y: 3, z: 2}
-  - {x: -1, y: -6, z: 2}
-  - {x: -1, y: -5, z: 1}
-  - {x: -1, y: -4, z: 0}
-  - {x: -1, y: -3, z: 0}
-  - {x: -1, y: -2, z: 0}
-  - {x: -1, y: -1, z: 0}
-  - {x: -1, y: 0, z: 0}
-  - {x: -1, y: 1, z: 0}
-  - {x: -1, y: 2, z: 1}
-  - {x: -1, y: 3, z: 2}
-  - {x: 0, y: -6, z: 2}
-  - {x: 0, y: -5, z: 1}
-  - {x: 0, y: -4, z: 0}
-  - {x: 0, y: -3, z: 0}
-  - {x: 0, y: -2, z: 0}
-  - {x: 0, y: -1, z: 0}
-  - {x: 0, y: 0, z: 0}
-  - {x: 0, y: 1, z: 0}
-  - {x: 0, y: 2, z: 1}
-  - {x: 0, y: 3, z: 2}
-  - {x: 1, y: -6, z: 2}
-  - {x: 1, y: -5, z: 1}
-  - {x: 1, y: -4, z: 0}
-  - {x: 1, y: -3, z: 0}
-  - {x: 1, y: -2, z: 0}
-  - {x: 1, y: -1, z: 0}
-  - {x: 1, y: 0, z: 0}
-  - {x: 1, y: 1, z: 0}
-  - {x: 1, y: 2, z: 1}
-  - {x: 1, y: 3, z: 2}
-  - {x: 2, y: -6, z: 2}
-  - {x: 2, y: -5, z: 1}
-  - {x: 2, y: -4, z: 0}
-  - {x: 2, y: -3, z: 0}
-  - {x: 2, y: -2, z: 0}
-  - {x: 2, y: -1, z: 0}
-  - {x: 2, y: 0, z: 0}
-  - {x: 2, y: 1, z: 0}
-  - {x: 2, y: 2, z: 1}
-  - {x: 2, y: 3, z: 2}
-  - {x: 3, y: -6, z: 2}
-  - {x: 3, y: -5, z: 1}
-  - {x: 3, y: -4, z: 1}
-  - {x: 3, y: -3, z: 1}
-  - {x: 3, y: -2, z: 1}
-  - {x: 3, y: -1, z: 1}
-  - {x: 3, y: 0, z: 1}
-  - {x: 3, y: 1, z: 1}
-  - {x: 3, y: 2, z: 1}
-  - {x: 3, y: 3, z: 2}
-  - {x: 4, y: -6, z: 2}
-  - {x: 4, y: -5, z: 2}
-  - {x: 4, y: -4, z: 2}
-  - {x: 4, y: -3, z: 2}
-  - {x: 4, y: -2, z: 2}
-  - {x: 4, y: -1, z: 2}
-  - {x: 4, y: 0, z: 2}
-  - {x: 4, y: 1, z: 2}
-  - {x: 4, y: 2, z: 2}
-  - {x: 4, y: 3, z: 2}
-  tileElevations: 02000000020000000200000002000000020000000200000002000000020000000200000002000000020000000100000001000000010000000100000001000000010000000100000001000000020000000200000001000000000000000000000000000000000000000000000000000000010000000200000002000000010000000000000000000000000000000000000000000000000000000100000002000000020000000100000000000000000000000000000000000000000000000000000001000000020000000200000001000000000000000000000000000000000000000000000000000000010000000200000002000000010000000000000000000000000000000000000000000000000000000100000002000000020000000100000000000000000000000000000000000000000000000000000001000000020000000200000001000000010000000100000001000000010000000100000001000000010000000200000002000000020000000200000002000000020000000200000002000000020000000200000002000000
-  dataLoaded: 1
-  totalTilesWritten: 100
-  lastOperationResult: 'Data validated: 100 tiles, elevation range [0, 2]'
-  minElevation: 0
-  maxElevation: 2
