--- Assets/_Project/Scripts/Water System/NewTerrainDataExample.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f77e9b3e4dfb1e747907abe258ac1fad
