--- Assets/_Project/Scripts/Water System/NewTerrainLoader.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bb5adb15a9e854d4d8507d03754e8eb5
