--- Assets/_Project/Scripts/Water System/Scriptable Object Scripts/NewTerrainData.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: de8d881e0a46a2542a9cb41813341c16
