--- Assets/_Project/Scripts/Water System/Scriptable Object Scripts/TerrainDataEditor.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d7214efdde31b8a49bd297484d226e2e
