+++ Assets/_Project/Scripts/Water System/TerrainData.asset
@@ -1,0 +1,21 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 86ae678e9f7c68f458c1bda6bf3225ba, type: 3}
+  m_Name: TerrainData
+  m_EditorClassIdentifier: 
+  tilePositions: []
+  tileElevations: 
+  dataLoaded: 0
+  totalTilesWritten: 0
+  lastOperationResult: No operation performed
+  minElevation: 2147483647
+  maxElevation: -2147483648
