--- Assets/_Project/Scripts/Water System/TileTypeSystem/TerrainLoader.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 65a641d034fd64a4da7ff57b1ba3c52f
